refactor(controller): extract browser/waveform/drag-drop/hotkeys/collections sub-controllers behind clear interfaces
