+++ src/egui_app/controller/browser_controller/delegates.rs
@@ -1,0 +1,259 @@
+use super::*;
+
+impl EguiController {
+    pub fn tag_browser_sample(&mut self, row: usize, tag: SampleTag) -> Result<(), String> {
+        self.browser().tag_browser_sample(row, tag)
+    }
+
+    pub fn tag_browser_samples(
+        &mut self,
+        rows: &[usize],
+        tag: SampleTag,
+        primary_visible_row: usize,
+    ) -> Result<(), String> {
+        self.browser()
+            .tag_browser_samples(rows, tag, primary_visible_row)
+    }
+
+    pub fn normalize_browser_sample(&mut self, row: usize) -> Result<(), String> {
+        self.browser().normalize_browser_sample(row)
+    }
+
+    pub fn normalize_browser_samples(&mut self, rows: &[usize]) -> Result<(), String> {
+        self.browser().normalize_browser_samples(rows)
+    }
+
+    pub fn rename_browser_sample(&mut self, row: usize, new_name: &str) -> Result<(), String> {
+        self.browser().rename_browser_sample(row, new_name)
+    }
+
+    pub fn delete_browser_sample(&mut self, row: usize) -> Result<(), String> {
+        self.browser().delete_browser_sample(row)
+    }
+
+    pub fn delete_browser_samples(&mut self, rows: &[usize]) -> Result<(), String> {
+        self.browser().delete_browser_samples(rows)
+    }
+
+    pub(in crate::egui_app::controller) fn resolve_browser_sample(
+        &self,
+        row: usize,
+    ) -> Result<helpers::TriageSampleContext, String> {
+        let source = self
+            .current_source()
+            .ok_or_else(|| "Select a source first".to_string())?;
+        let index = self
+            .visible_browser_indices()
+            .get(row)
+            .copied()
+            .ok_or_else(|| "Sample not found".to_string())?;
+        let entry = self
+            .wav_entries
+            .get(index)
+            .cloned()
+            .ok_or_else(|| "Sample not found".to_string())?;
+        let absolute_path = source.root.join(&entry.relative_path);
+        Ok(helpers::TriageSampleContext {
+            source,
+            entry,
+            absolute_path,
+        })
+    }
+
+    pub(in crate::egui_app::controller) fn prune_cached_sample(
+        &mut self,
+        source: &SampleSource,
+        relative_path: &Path,
+    ) {
+        if let Some(cache) = self.wav_cache.get_mut(&source.id) {
+            cache.retain(|entry| entry.relative_path != relative_path);
+        }
+        if self.selected_source.as_ref() == Some(&source.id) {
+            self.wav_entries
+                .retain(|entry| entry.relative_path != relative_path);
+            self.rebuild_wav_lookup();
+            self.rebuild_browser_lists();
+            self.label_cache
+                .insert(source.id.clone(), self.build_label_cache(&self.wav_entries));
+        } else {
+            self.label_cache.remove(&source.id);
+        }
+        self.rebuild_missing_lookup_for_source(&source.id);
+        self.clear_loaded_sample_if(source, relative_path);
+    }
+
+    pub(in crate::egui_app::controller) fn clear_loaded_sample_if(
+        &mut self,
+        source: &SampleSource,
+        relative_path: &Path,
+    ) {
+        self.invalidate_cached_audio(&source.id, relative_path);
+        if self.selected_source.as_ref() == Some(&source.id) {
+            if self.selected_wav.as_deref() == Some(relative_path) {
+                self.selected_wav = None;
+            }
+            if self.loaded_wav.as_deref() == Some(relative_path) {
+                self.loaded_wav = None;
+            }
+            if self.ui.loaded_wav.as_deref() == Some(relative_path) {
+                self.ui.loaded_wav = None;
+            }
+        }
+        if let Some(audio) = self.loaded_audio.as_ref()
+            && audio.source_id == source.id
+            && audio.relative_path == relative_path
+        {
+            self.loaded_audio = None;
+            self.decoded_waveform = None;
+            self.ui.waveform.image = None;
+            self.ui.waveform.playhead = PlayheadState::default();
+            self.ui.waveform.selection = None;
+            self.ui.waveform.selection_duration = None;
+            self.selection.clear();
+        }
+    }
+
+    pub(in crate::egui_app::controller) fn refresh_waveform_for_sample(
+        &mut self,
+        source: &SampleSource,
+        relative_path: &Path,
+    ) {
+        self.invalidate_cached_audio(&source.id, relative_path);
+        let loaded_matches = self.loaded_audio.as_ref().is_some_and(|audio| {
+            audio.source_id == source.id && audio.relative_path == relative_path
+        });
+        let selected_matches = self.selected_source.as_ref() == Some(&source.id)
+            && self.selected_wav.as_deref() == Some(relative_path);
+        if selected_matches || loaded_matches {
+            self.loaded_wav = None;
+            self.ui.loaded_wav = None;
+            if let Err(err) = self.load_waveform_for_selection(source, relative_path) {
+                self.set_status(err, StatusTone::Warning);
+            }
+        }
+    }
+
+    pub(in crate::egui_app::controller) fn reexport_collections_for_sample(
+        &mut self,
+        source_id: &SourceId,
+        relative_path: &Path,
+    ) {
+        let mut targets = Vec::new();
+        for collection in self.collections.iter() {
+            if collection
+                .members
+                .iter()
+                .any(|m| &m.source_id == source_id && m.relative_path == relative_path)
+            {
+                targets.push((
+                    collection.id.clone(),
+                    collection_export::resolved_export_dir(
+                        collection,
+                        self.collection_export_root.as_deref(),
+                    ),
+                ));
+            }
+        }
+        let member = CollectionMember {
+            source_id: source_id.clone(),
+            relative_path: relative_path.to_path_buf(),
+        };
+        for (collection_id, export_dir) in targets {
+            collection_export::delete_exported_file(export_dir.clone(), &member);
+            if let Err(err) = self.export_member_if_needed(&collection_id, &member) {
+                self.set_status(err, StatusTone::Warning);
+            }
+        }
+    }
+
+    pub(in crate::egui_app::controller) fn update_collections_for_rename(
+        &mut self,
+        source_id: &SourceId,
+        old_relative: &Path,
+        new_relative: &Path,
+    ) -> bool {
+        let mut changed = false;
+        let mut exports: Vec<(CollectionId, Option<PathBuf>)> = Vec::new();
+        for collection in self.collections.iter_mut() {
+            let mut touched = false;
+            for member in collection.members.iter_mut() {
+                if &member.source_id == source_id && member.relative_path == old_relative {
+                    member.relative_path = new_relative.to_path_buf();
+                    touched = true;
+                    changed = true;
+                }
+            }
+            if touched {
+                exports.push((
+                    collection.id.clone(),
+                    collection_export::resolved_export_dir(
+                        collection,
+                        self.collection_export_root.as_deref(),
+                    ),
+                ));
+            }
+        }
+        if changed {
+            let member = CollectionMember {
+                source_id: source_id.clone(),
+                relative_path: new_relative.to_path_buf(),
+            };
+            for (collection_id, export_dir) in exports {
+                let old_member = CollectionMember {
+                    source_id: source_id.clone(),
+                    relative_path: old_relative.to_path_buf(),
+                };
+                collection_export::delete_exported_file(export_dir.clone(), &old_member);
+                if let Err(err) = self.export_member_if_needed(&collection_id, &member) {
+                    self.set_status(err, StatusTone::Warning);
+                }
+            }
+            self.refresh_collections_ui();
+        }
+        changed
+    }
+
+    pub(in crate::egui_app::controller) fn remove_sample_from_collections(
+        &mut self,
+        source_id: &SourceId,
+        relative_path: &Path,
+    ) -> bool {
+        let mut changed = false;
+        for collection in self.collections.iter_mut() {
+            let member = CollectionMember {
+                source_id: source_id.clone(),
+                relative_path: relative_path.to_path_buf(),
+            };
+            if collection.remove_member(source_id, &member.relative_path) {
+                changed = true;
+                let export_dir = collection_export::resolved_export_dir(
+                    collection,
+                    self.collection_export_root.as_deref(),
+                );
+                collection_export::delete_exported_file(export_dir, &member);
+            }
+        }
+        if changed {
+            self.refresh_collections_ui();
+        }
+        changed
+    }
+
+    pub(in crate::egui_app::controller) fn refocus_after_filtered_removal(
+        &mut self,
+        primary_visible_row: usize,
+    ) {
+        if matches!(self.ui.browser.filter, TriageFlagFilter::All) {
+            return;
+        }
+        if self.ui.browser.visible.is_empty() || self.ui.browser.selected_visible.is_some() {
+            return;
+        }
+        if self.random_navigation_mode_enabled() {
+            self.focus_random_visible_sample();
+            return;
+        }
+        let target_row = primary_visible_row.min(self.ui.browser.visible.len().saturating_sub(1));
+        self.focus_browser_row_only(target_row);
+    }
+}
