+++ src/egui_app/controller/browser_controller/mod.rs
@@ -1,0 +1,13 @@
+mod actions;
+mod delegates;
+mod helpers;
+
+pub(crate) use actions::BrowserActions;
+pub(crate) use helpers::BrowserController;
+
+use super::collection_export;
+use super::collection_items_helpers::file_metadata;
+use super::*;
+use crate::sample_sources::collections::CollectionMember;
+use std::path::{Path, PathBuf};
+
